add shared examples and create base model to inherited from other models
