+++ spec/lib/author_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 require_relative '../../lib/alchemy_language/generator/alchemy_request'
-require_relative '../../lib/alchemy_language/active_method/extra'
+require_relative '../../lib/alchemy_language/active_model/base'
 require_relative '../../lib/alchemy_language/model/author'
 
 describe AlchemyLanguage::Author do
@@ -11,21 +11,7 @@ describe AlchemyLanguage::Author do
     allow(RestClient).to receive("get").and_return(json_response("author.json"))
   end
 
-  it "has accessor for path" do
-    is_expected.to respond_to(:path)
-  end
-
-  it "has accessor for auth_token" do
-    is_expected.to respond_to(:auth_token)
-  end
-
-  it "has accessor for type" do
-    is_expected.to respond_to(:type)
-  end
-
-  it "has accessor for json_result" do
-    is_expected.to respond_to(:json_result)
-  end
+  it_should_behave_like "model_accessor"
 
   describe "#add_method" do
     it "author class has method called status and return result from json_result" do
